+++ tests/DBMigratorTest.php
@@ -30,12 +30,12 @@ final class DBMigratorTest extends TestCase
         $this->assertStringContainsString($GLOBALS['wpdb']->prefix . 'matchmaker_logs', $GLOBALS['__mm_dbdelta_sql']);
         $this->assertStringContainsString('has_one_on_one', $GLOBALS['__mm_dbdelta_sql']);
         $this->assertStringContainsString('is_parent_applying', $GLOBALS['__mm_dbdelta_sql']);
-        $this->assertEquals('2.7.0', $GLOBALS['__mm_options']['mm_matchmaking_db_v2_version']);
+        $this->assertEquals('2.8.0', $GLOBALS['__mm_options']['mm_matchmaking_db_v2_version']);
     }
 
     public function test_maybe_migrate_skips_when_already_installed(): void
     {
-        $GLOBALS['__mm_options']['mm_matchmaking_db_v2_version'] = '2.7.0';
+        $GLOBALS['__mm_options']['mm_matchmaking_db_v2_version'] = '2.8.0';
         $migrator = DBMigrator::instance();
         $migrator->maybe_migrate();